+++ crates/unleash-edge-types/src/errors.rs
@@ -123,7 +123,9 @@ pub enum EdgeError {
     FrontendExpectedToBeHydrated(String),
     FrontendNotYetHydrated(FrontendHydrationMissing),
     HealthCheckError(String),
+    HeartbeatError(String, StatusCode),
     InvalidBackupFile(String, String),
+    InvalidLicense(String),
     InvalidServerUrl(String),
     InvalidEtag,
     InvalidToken,
@@ -184,6 +186,7 @@ impl Display for EdgeError {
                 write!(f, "Failed to build cert {cert_error:?}")
             }
             EdgeError::ClientBuildError(e) => write!(f, "Failed to build client {e:?}"),
+            EdgeError::InvalidLicense(msg) => write!(f, "License is invalid: [{msg}]"),
             EdgeError::InvalidServerUrl(msg) => write!(f, "Failed to parse server url: [{msg}]"),
             EdgeError::EdgeTokenError => write!(f, "Edge token error"),
             EdgeError::EdgeTokenParseError => write!(f, "Failed to parse token response"),
@@ -205,6 +208,12 @@ impl Display for EdgeError {
             }
             EdgeError::HealthCheckError(message) => {
                 write!(f, "{message}")
+            }
+            EdgeError::HeartbeatError(message, status_code) => {
+                write!(
+                    f,
+                    "Received status code {status_code} when trying to send heartbeat to upstream server: {message}"
+                )
             }
             EdgeError::ReadyCheckError(message) => {
                 write!(f, "{message}")
@@ -267,7 +276,12 @@ impl IntoResponse for EdgeError {
                 "access": frontend_hydration_missing.clone()
             }).to_string())),
             EdgeError::HealthCheckError(_) => Response::builder().status(self.status_code()).body(Body::empty()),
+            EdgeError::HeartbeatError(message, status_code) => Response::builder().status(self.status_code()).body(Body::from(json!({
+                "explanation": format!("Received a non 200 status code when trying to validate token upstream: {message}"),
+                "status_code": status_code.as_str()
+            }).to_string())),
             EdgeError::InvalidBackupFile(_, _) => Response::builder().status(self.status_code()).body(Body::empty()),
+            EdgeError::InvalidLicense(_) => Response::builder().status(self.status_code()).body(Body::empty()),
             EdgeError::InvalidServerUrl(_) => Response::builder().status(self.status_code()).body(Body::empty()),
             EdgeError::JsonParseError(_) => Response::builder().status(self.status_code()).body(Body::empty()),
             EdgeError::NoFeaturesFile => Response::builder().status(self.status_code()).body(Body::empty()),
@@ -306,6 +320,7 @@ impl EdgeError {
             EdgeError::ClientBuildError(_) => StatusCode::INTERNAL_SERVER_ERROR,
             EdgeError::ClientFeaturesParseError(_) => StatusCode::INTERNAL_SERVER_ERROR,
             EdgeError::ClientFeaturesFetchError(_) => StatusCode::INTERNAL_SERVER_ERROR,
+            EdgeError::InvalidLicense(_) => StatusCode::PAYMENT_REQUIRED,
             EdgeError::InvalidServerUrl(_) => StatusCode::INTERNAL_SERVER_ERROR,
             EdgeError::PersistenceError(_) => StatusCode::INTERNAL_SERVER_ERROR,
             EdgeError::JsonParseError(_) => StatusCode::INTERNAL_SERVER_ERROR,
@@ -323,6 +338,7 @@ impl EdgeError {
                 StatusCode::from_u16(status_code.as_u16()).unwrap()
             }
             EdgeError::HealthCheckError(_) => StatusCode::INTERNAL_SERVER_ERROR,
+            EdgeError::HeartbeatError(_, status_code) => *status_code,
             EdgeError::ReadyCheckError(_) => StatusCode::INTERNAL_SERVER_ERROR,
             EdgeError::ClientHydrationFailed(_) => StatusCode::INTERNAL_SERVER_ERROR,
             EdgeError::ClientCacheError => StatusCode::INTERNAL_SERVER_ERROR,