+++ crates/unleash-edge/src/middleware/mod.rs
@@ -1,6 +1,6 @@
 pub mod allow_list;
+pub mod client_metrics;
 pub mod consumption;
 pub mod deny_list;
 pub mod etag;
-pub mod log_request;
 pub mod validate_token;